requirements_news.txt
streamlit==1.19.0
google-search-python==1.1.0
scikit-learn==1.1.2
requests==2.28.1
pandas==1.5.3
